refactored the Start Menu into it's own file. Everything is still working as was
